requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pydantic
httpx
jinja2
python-multipart
pywebview
